+++ docs/source/changelog.rst
@@ -5,6 +5,11 @@ Changelog
 
 .. warning:: Please follow :ref:`upgrading-kegbot` for general upgrade steps.
 
+Current Version (in development)
+--------------------------------
+
+* Fixed issue #128 (bug while ending a keg)
+
 Version 0.9.9 (2013-05-14)
 --------------------------